+++ CHANGELOG.md
@@ -15,6 +15,7 @@ For instance:
 
 ### Changed
 - [Gutenberg] Compiler: The raw block data is now appended on the scanner data to improve the class name detection [#53](https://github.com/wind-press/windpress/issues/53)
+- [Wizard] The Wizard's data is now can saved directly without the need to switch the Files tab first.
 
 ### Fixed
 - [Bricks] Plain Classes feature doesn't work correctly with Bricks' templates